new_portfolio/requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
Flask-Admin==1.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.3
python-dotenv==1.0.1
Werkzeug==3.0.2
WTForms==3.1.2
